requirements.txt
matplotlib~=3.3.4
numpy>=1.16.1
Pillow~=8.1.0
tensorflow~=2.3.2